+++ app/src/main/java/com/farware/recipesaver/feature_recipe/domain/model/recipe/RecipeIngredient.kt
@@ -5,7 +5,7 @@ import java.math.RoundingMode
 data class RecipeIngredient(
     val recipeIngredientId: Long? = null,
     val recipeId: Long,
-    val ingredientId: Int,
+    val ingredientId: Long,
     val measureId: Int,
     val amount: String,
     val measure: String,
