+++ app/src/main/java/com/farware/recipesaver/feature_recipe/domain/model/recipe/relations/RecipeIngredientWithIngredient.kt
@@ -3,7 +3,7 @@ package com.farware.recipesaver.feature_recipe.domain.model.recipe.relations
 data class RecipeIngredientWithIngredient(
     val recipeIngredientId: Long? = null,
     val recipeId: Long,
-    val ingredientId: Int,
+    val ingredientId: Long,
     val measureId: Int,
     val amount: String,
     val measure: String,
